feat: Cart 수량변경,삭제,조회 로직이랑 장바구니 담기 누르면 order로 연결 기능 초안 완료

+++ src/main/java/com/elice/kittyandpuppy/module/order/service/OrderItemService.java
@@ -58,7 +58,7 @@ public class OrderItemService {
     }
 
     @Transactional
-    public OrderItem update(Long id, Long amount) {
+    public OrderItem update(Long id, int amount) {
         OrderItem orderItem = orderItemRepository.findById(id).orElseThrow();
         orderItem.setProductAmount(amount);
         return orderItemRepository.save(orderItem);
